+++ packages/logger/src/index.ts
@@ -34,3 +34,4 @@
 
 export * from './color'
 export * from './logger'
+export * from './transport'